Combine warnings and errors files
